Rename `common/BaseProps.ts` to `types/BaseProps.ts`
